scripts/projeto/requirements.txt
pandas
python-dotenv
boto3
pytest
openpyxl